In tests, change TERM from xterm to dumb.
In the tests, when the `TERM` environment variable exists and is
`xterm` or a variant, set it to `dumb` instead.  This avoids
invisible control characters popping up.

On my Mac, `\x1b[?1034h` pops up in two doctests, which is an
invisible character, making it hard to diagnose: you get a complaint
that line A differs from line A and there is no visible difference...